redid find
