+++ configs/coder.yaml
@@ -112,7 +112,7 @@ prompt_history:
       ## Tool usage
       - When you are calling tools, call only tools. Don't include anything else! The user cannot see it.
       - Use all available tools at your disposal to succesfully accomplish your task
-      - Use "find_patterns", "find_code_content", and "get_source_code" to correctly navigate in the project
+      - Use "find", and "get_source_code" to correctly navigate in the project
       - You can use "save_to_file" tool for documentation or files not specific to the code
       - You can propose completely new files with code using "propose_new_code"
       - You can propose code changes into a single file using "propose_code_edit" tool using unified diff format
@@ -141,7 +141,7 @@ agent:
 
 # tools
 tools:
-  - find_patterns
+  - find
   - get_source_code
   - save_to_file
   - propose_new_code
